Move RouterContextProvider to shared
