+++ app/src/editor/canvas/components/CanvasController/index.jsx
@@ -2,7 +2,7 @@ import React, { useContext, useEffect, useMemo, useCallback, useState } from 're
 import { Helmet } from 'react-helmet'
 import LoadingIndicator from '$userpages/components/LoadingIndicator'
 
-import * as RouterContext from '$editor/shared/components/RouterContext'
+import { RouterContextProvider, RouterContext } from '$shared/components/RouterContextProvider'
 import usePending, { useAnyPending, Provider as PendingProvider } from '$editor/shared/hooks/usePending'
 import { Provider as PermissionsProvider } from '$editor/canvas/hooks/useCanvasPermissions'
 
@@ -20,7 +20,7 @@ import styles from './CanvasController.pcss'
 function useCanvasLoadEffect() {
     const canvas = useCanvas()
     const load = useCanvasLoadCallback()
-    const { match } = useContext(RouterContext.Context)
+    const { match } = useContext(RouterContext)
     const { isPending } = usePending('LOAD')
 
     const { id: urlId } = match.params
@@ -89,7 +89,7 @@ export function useController() {
 }
 
 function useCanvasCreateEffect() {
-    const { match } = useContext(RouterContext.Context)
+    const { match } = useContext(RouterContext)
     const { isPending } = usePending('CREATE')
 
     const create = useCanvasCreateCallback()
@@ -125,7 +125,7 @@ function CanvasLoadingIndicator() {
 }
 
 const CanvasControllerProvider = ({ children }) => (
-    <RouterContext.Provider>
+    <RouterContextProvider>
         <PendingProvider>
             <PermissionsProvider>
                 <CanvasLoadingIndicator />
@@ -133,7 +133,7 @@ const CanvasControllerProvider = ({ children }) => (
                 {children || null}
             </PermissionsProvider>
         </PendingProvider>
-    </RouterContext.Provider>
+    </RouterContextProvider>
 )
 
 export { CanvasControllerProvider as Provider }
