+++ app/src/editor/canvas/components/CanvasController/useCanvasCreateCallback.jsx
@@ -1,14 +1,14 @@
 import { useContext, useCallback } from 'react'
 
 import useIsMountedRef from '$shared/utils/useIsMountedRef'
-import * as RouterContext from '$editor/shared/components/RouterContext'
+import { RouterContext } from '$shared/components/RouterContextProvider'
 import usePending from '$editor/shared/hooks/usePending'
 
 import links from '../../../../links'
 import * as services from '../../services'
 
 export default function useCanvasCreateCallback() {
-    const { history } = useContext(RouterContext.Context)
+    const { history } = useContext(RouterContext)
     const { isPending, wrap } = usePending('CREATE')
     const isMountedRef = useIsMountedRef()
