+++ app/src/editor/canvas/components/CanvasController/useCanvasDuplicateCallback.jsx
@@ -1,14 +1,14 @@
 import { useContext, useCallback } from 'react'
 
 import useIsMountedRef from '$shared/utils/useIsMountedRef'
-import * as RouterContext from '$editor/shared/components/RouterContext'
+import { RouterContext } from '$shared/components/RouterContextProvider'
 import usePending from '$editor/shared/hooks/usePending'
 
 import links from '../../../../links'
 import * as services from '../../services'
 
 export default function useCanvasDuplicateCallback() {
-    const { history } = useContext(RouterContext.Context)
+    const { history } = useContext(RouterContext)
     const { isPending, wrap } = usePending('DUPLICATE')
     const isMountedRef = useIsMountedRef()
