+++ app/src/shared/contexts/Router/index.js
@@ -1,0 +1,5 @@
+// @flow
+
+import React, { type Context } from 'react'
+
+export default (React.createContext({}): Context<{}>)
